fix: middleware build error
